Added the original optional ge stage
I noticed that the optional gold edition stage from ocp was not using the original stage used in the ge from xbox 360, it was just using the stage that comes from ocp/ui/models/m_team_stage.igb. This previous stage looks odd with the ge mannequins because it looks like everyone is slightly floating above the ground but the new stage fixes this issue. Another issue I noticed with the previous stage is that some characters are not fitted on the stage properly, such as Luke Cages and Colossus's  legs would be clipped inside of the stairs of the stage. This has also been fixed with the new stage added. Same thing applies for Dr. Stranges cape even though the xbox 360 ge stage still has a small amount of clipping with his cape, its better than the clipping with the previous stage.

I added a preview folder found in the optionals/ge stage folder to see the comparisons.