Few mannequin adjustments for ge stage
Mostly changed some mannequins to be either higher or lower so their feet properly align with the ground or match the xbox 360 original ge stage.

Moved ghost rider slightly forward to avoid clipping surfers surfboard.

Dr strange still needs to be brought forward a small amount in order for his cape to avoid clipping the stairs.

The Thing needs to be lowered a small amount so his feet touch the floor.