Added testguy animation file, which some NPCs use
It's been reported that some NPCs end up in T-pose, probably when a limit is reached. Adding the missing animation set seems to fix that.
There are other NPCs which have a non-existent animation set referenced in NPCstat, but I suggest to leave them alone until an issue is reported and it's reported to be fixed by adding an animation set.