Removed useless events fr. Hawkeye PS
(again: stupid mistake I made. I have 161 changes in progress, makes it really hard to make other commits, sorry.)

Lars changed the events to actual skinsegment triggers, akin to how it is with Deadpool's PS. Events are useless now.
There are a few differences to the original though.
Might be worth re-visiting, as it might be possible to reduce the trigger count by combining some of the skinsegments.